requirements.txt
scikit-learn==1.6.1
pandas==2.2.2
numpy==2.0.2
xgboost==3.0.2
joblib==1.5.1
gradio==5.31.0
google-genai==1.24.0
python-dotenv
category_encoders